requirements.txt
Django
djangorestframework
django-cors-headers
psycopg2-binary
djangorestframework-simplejwt
django-allauth
Pillow
python-decouple
django-filter
django-extensions